requirements.txt
pandas
numpy==1.26.4
scikit-learn
xgboost==2.0.3
imbalanced-learn
joblib
fastapi[standard]
streamlit
shap
evidently==0.4.19
matplotlib
seaborn
jupyter
sqlalchemy
psycopg2-binary
python-dotenv
mlflow
dagshub